archdiag-iac-agents/requirements.txt
# Azure AI Agent Framework (PREVIEW - requires --pre flag)
# Install with: pip install -r requirements.txt
agent-framework-azure-ai>=1.0.0b251016

# Azure SDK - Computer Vision
azure-ai-vision-imageanalysis>=1.0.0
azure-cognitiveservices-vision-computervision>=0.9.0

# Azure SDK - Identity & Management (Managed Identity preferred)
azure-identity>=1.16.0
azure-mgmt-resource>=23.0.0
azure-mgmt-cognitiveservices>=13.5.0
azure-mgmt-subscription>=3.1.0

# Azure SDK - AI Foundry
azure-ai-inference>=1.0.0

# Configuration & Environment
pydantic>=2.8.0
pydantic-settings>=2.3.0
python-dotenv>=1.0.0

# Web Framework
fastapi>=0.115.0
uvicorn>=0.31.0
python-multipart>=0.0.9

# HTTP & Async
httpx>=0.27.0
aiofiles>=24.1.0

# Image Processing
Pillow>=10.4.0

# Data Processing
pydantic>=2.8.0
typing-extensions>=4.12.0

# Logging & Monitoring
azure-monitor-opentelemetry>=1.6.0
opentelemetry-api>=1.25.0
opentelemetry-sdk>=1.25.0

# Utilities
tenacity>=9.0.0
jinja2>=3.1.4

# Development tools (optional)
pytest>=8.3.0
pytest-asyncio>=0.23.0
black>=24.8.0
ruff>=0.6.0
